begin to update mongodb
