+++ client/public/index.html
@@ -26,6 +26,8 @@
   <script>
     function onSuccess(googleUser) {
       console.log('Logged in as: ' + googleUser.getBasicProfile().getName());
+      
+      // ******* NEED REDIRECT CODE TO SEND TO PAGE FOR SIGN UP *************
     }
     function onFailure(error) {
       console.log(error);
