+++ client/src/utils/API.js
@@ -2,7 +2,7 @@ import axios from "axios";
 
 export default {
   // Gets all books
-  getBooks: function() {
+  getNudges: function() {
     return axios.get("/api/books");
   },
   // Gets the book with the given id
